lesson-09-long-term-agent-memory/demo/requirements.txt
google-adk>=1.17.0
google-cloud-aiplatform>=1.124.0
